refac:  commands module from hello-world extension
