+++ extensions/hello-world/src/index.js
@@ -1,5 +1,5 @@
-import toolbarModule from './toolbarModule.js';
-import MessageDialog from './components/MessageDialog';
+import toolbarModule from './toolbarModule';
+import commandsModule from './commandsModule';
 
 /**
  *
@@ -37,51 +37,7 @@ export default {
     return toolbarModule;
   },
   getCommandsModule({ servicesManager }) {
-    const { UIDialogService } = servicesManager.services;
-
-    const showHelloWorldDialog = (title, content) => {
-      if (!UIDialogService) {
-        console.warn('Unable to show dialog; no UI Dialog Service available.');
-        return;
-      }
-
-      UIDialogService.dismiss({ id: 'hello-world' });
-      UIDialogService.create({
-        id: 'hello-world',
-        centralize: true,
-        isDraggable: false,
-        showOverlay: true,
-        content: MessageDialog,
-        contentProps: {
-          messageTitle: title,
-          messageContent: content,
-          dialogCloseCallback: () =>
-            UIDialogService.dismiss({ id: 'hello-world' }),
-        },
-      });
-    };
-
-    const actions = {
-      // Store Contexts + Options
-      helloWorld: ({ title, content }) => {
-        showHelloWorldDialog(title, content);
-      },
-    };
-
-    return {
-      actions,
-      definitions: {
-        helloWorld: {
-          commandFn: actions.helloWorld,
-          storeContexts: ['viewports'],
-          options: {
-            title: 'Hello World!',
-            content:
-              'This is the Exercise 1 that was proposed by Igor Octaviano and was done by Vinícius Alves.',
-          },
-        },
-      },
-    };
+    return commandsModule({ servicesManager });
   },
 };
